@@ -25,6 +25,7 @@
   },
   "license": "AGPL",
   "devDependencies": {
+    "@svgr/cli": "^5.5.0",
     "@types/asn1js": "0.0.1",
     "@types/classnames": "^2.2.10",
     "@types/extract-zip": "^1.6.2",
@@ -41,6 +42,7 @@
     "@types/xmldom": "^0.1.30",
     "@typescript-eslint/eslint-plugin": "^2.34.0",
     "awesome-typescript-loader": "^5.2.1",
+    "bootstrap": "^5.0.1",
     "colors": "^1.4.0",
     "cross-env": "^7.0.2",
     "css-loader": "^5.2.6",
@@ -53,7 +55,6 @@
     "eslint-plugin-react": "^7.20.3",
     "eslint-plugin-react-hooks": "^3.0.0",
     "extract-zip": "^1.7.0",
-    "fortify-prepare": "github:PeculiarVentures/fortify-prepare",
     "json-parser": "^3.1.2",
     "mocha": "^7.2.0",
     "node-gyp": "^7.1.2",
@@ -64,9 +65,10 @@
     "sass-loader": "^10.2.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^1.2.1",
+    "svg-url-loader": "^7.1.1",
     "ts-node": "^8.10.2",
     "typescript": "^3.9.7",
-    "webpack": "^4.43.0",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.11",
     "webpack-merge": "^4.2.2"
   },
@@ -82,7 +84,6 @@
     "@webcrypto-local/cards": "^1.6.3",
     "@webcrypto-local/server": "^1.6.4",
     "asn1js": "^2.0.26",
-    "bootstrap": "^5.0.1",
     "classnames": "^2.2.6",
     "get-proxy-settings": "^0.1.11",
     "jose-jwe-jws": "github:microshine/js-jose",
@@ -99,7 +100,6 @@
     "semver": "^6.3.0",
     "sudo-prompt": "^9.2.1",
     "tsyringe": "^4.3.0",
-    "typeface-montserrat": "^1.1.13",
     "uuid": "^8.3.1",
     "webcrypto-core": "^1.1.8",
     "winston": "^3.3.3",
